gralloc4: Only free allocated handles in error path
If we encounter an error while allocating handles, some error paths
could end up trying to freeing more handles that we allocated so far.

Also add some extra checks and logging in mali_grallic_ion_allocate() to
catch incorrect calls to this function, and ensure the array of
filedescriptors in handles is initialized.

Bug: 241512108
Test: v2/android-platinum/health/unit/camera
Change-Id: Idd2b09ed11424b110d485e10656771aafe0b20a9